accessible/tests/mochitest/treeupdate/test_menu.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL menu hierarchy tests">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Invokers

    function openMenu(aID)
    {
      this.menuNode = getNode(aID),

      this.eventSeq = [
        new invokerChecker(EVENT_FOCUS, this.menuNode)
      ];

      this.invoke = function openMenu_invoke()
      {
        var tree;
        if (LINUX || SOLARIS) {
          tree =
            { PARENT_MENUITEM: [ ] };

        } else {
          tree =
            { PARENT_MENUITEM: [
              { MENUPOPUP: [ ] }
            ] };
        }
        testAccessibleTree(aID, tree);

        // Show menu.
        this.menuNode.open = true;
      }

      this.finalCheck = function openMenu_finalCheck()
      {
        var tree;
        if (LINUX || SOLARIS) {
          tree =
            { PARENT_MENUITEM: [
              { MENUITEM: [ ] },
              { MENUITEM: [ ] }
            ] };

        } else {
          tree =
            { PARENT_MENUITEM: [
              { MENUPOPUP: [
                { MENUITEM: [ ] },
                { MENUITEM: [ ] }
              ] }
            ] };
        }
        testAccessibleTree(aID, tree);
      }

      this.getID = function openMenu_getID()
      {
        return "open menu " + prettyName(aID);
      }
    }

    ////////////////////////////////////////////////////////////////////////////
    // Test

    var gQueue = null;
    function doTest()
    {
      gQueue = new eventQueue();
      gQueue.push(new openMenu("menu"));
      gQueue.invoke(); // Will call SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=249292"
         title="Ensure accessible children for toolbarbutton types 'menu' and 'menu-button'">
        Mozilla Bug 249292
      </a>
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=630486"
         title="Don't force accessible creation for popup children.">
        Mozilla Bug 630486
      </a>
      <br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <menubar>
        <menu id="menu" label="menu">
          <menupopup>
            <menuitem label="menuitem"/>
            <menuitem label="menuitem"/>
          </menupopup>
        </menu>
      </menubar>
    </vbox>
  </hbox>

</window>
